+++ firebase-vscode/src/core/emulators.ts
@@ -6,6 +6,7 @@ import { EmulatorsStatus, RunningEmulatorInfo } from "../messaging/types";
 import { EmulatorHubClient } from "../../../src/emulator/hubClient";
 import { GetEmulatorsResponse } from "../../../src/emulator/hub";
 import { EmulatorInfo } from "../emulator/types";
+import { getSettings } from "../utils/settings";
 export class EmulatorsController implements Disposable {
   constructor(private broker: ExtensionBrokerImpl) {
     this.emulatorStatusItem.command = "firebase.openFirebaseRc";
@@ -21,6 +22,25 @@ export class EmulatorsController implements Disposable {
         this.setEmulatorsStarting();
       }),
     );
+
+    // Subscription to open up settings window
+    this.subscriptions.push(
+      broker.on("fdc.open-emulator-settings", () => {
+        vscode.commands.executeCommand( 'workbench.action.openSettings', 'firebase.emulators' );
+      })
+    );
+
+    // Subscription to trigger clear emulator data when button is clicked.
+    this.subscriptions.push(
+      broker.on("fdc.clear-emulator-data", () => {
+        vscode.commands.executeCommand("firebase.emulators.clearData");
+      }),
+    );
+
+    // Subscription to trigger emulator exports when button is clicked.
+    this.subscriptions.push(broker.on("runEmulatorsExport", () => {
+      vscode.commands.executeCommand("firebase.emulators.exportData")
+    }));
   }
 
   readonly emulatorStatusItem = vscode.window.createStatusBarItem("emulators");
@@ -37,6 +57,17 @@ export class EmulatorsController implements Disposable {
   private readonly emulatorsStoppped = vscode.commands.registerCommand(
     "firebase.emulators.stopped",
     this.setEmulatorsStopped.bind(this),
+  );
+
+  private readonly clearEmulatorDataCommand = vscode.commands.registerCommand(
+    "firebase.emulators.clearData",
+    this.clearDataConnectData.bind(this),
+  );
+
+
+  private readonly exportEmulatorDataCommand = vscode.commands.registerCommand(
+    "firebase.emulators.exportData",
+    this.exportEmulatorData.bind(this),
   );
 
   readonly emulators: { status: EmulatorsStatus; infos?: RunningEmulatorInfo } =
@@ -107,11 +138,8 @@ export class EmulatorsController implements Disposable {
   async findRunningCliEmulators(): Promise<
     { status: EmulatorsStatus; infos?: RunningEmulatorInfo } | undefined
   > {
-    const projectId = firebaseRC.value?.tryReadValue?.projects?.default;
-    // TODO: think about what to without projectID, in potentially a logged out mode
-    const hubClient = new EmulatorHubClient(projectId!);
-
-    if (hubClient.foundHub()) {
+    const hubClient = this.getHubClient();
+    if (hubClient) {
       const response: GetEmulatorsResponse = await hubClient.getEmulators();
 
       if (Object.values(response)) {
@@ -119,11 +147,38 @@ export class EmulatorsController implements Disposable {
       } else {
         this.setEmulatorsStopped();
       }
+    }
+    return this.emulators;
+  }
+
+  async clearDataConnectData(): Promise<void> {
+    const hubClient = this.getHubClient();
+    if (hubClient) {
+      await hubClient.clearDataConnectData();
+      vscode.window.showInformationMessage(`Data Connect emulator data has been cleared.`);
+    }
+  }
+
+  async exportEmulatorData(): Promise<void> {
+    const settings = getSettings();
+    const exportDir = settings.exportPath;
+    const hubClient = this.getHubClient();
+    if (hubClient) {
+      // TODO: Make exportDir configurable
+      await hubClient.postExport({path: exportDir, initiatedBy: "Data Connect VSCode extension"});
+      vscode.window.showInformationMessage(`Emulator Data exported to ${exportDir}`);
+    }
+  }
+
+  private getHubClient(): EmulatorHubClient | undefined {
+    const projectId = firebaseRC.value?.tryReadValue?.projects?.default;
+    // TODO: think about what to without projectID, in potentially a logged out mode
+    const hubClient = new EmulatorHubClient(projectId!);
+    if (hubClient.foundHub()) {
+      return hubClient;
     } else {
       this.setEmulatorsStopped();
     }
-
-    return this.emulators;
   }
 
   public areEmulatorsRunning() {
@@ -135,5 +190,7 @@ export class EmulatorsController implements Disposable {
     this.findRunningEmulatorsCommand.dispose();
     this.emulatorStatusItem.dispose();
     this.emulatorsStoppped.dispose();
+    this.clearEmulatorDataCommand.dispose();
+    this.exportEmulatorDataCommand.dispose();
   }
 }